requirements.txt
instagrapi
pandas
python-dotenv 


